Moved temp.js methods over to `campaign.js`
